update zoom tool

+++ sumpdf/src/PageView.hh
@@ -74,7 +74,7 @@ private :
 	
 	// dragging
 	QPointF		m_start_drag ;
-	int			m_page_size ;
+	double			m_page_size ;
 } ;
 
 } // end of namespace
